[refactor] api/us/country/today route
