+++ queries.js
@@ -6,6 +6,22 @@ var async = require('async');
 var _ = require('underscore');
 
 /*************************************************************
+Add Underscore Mixin to sort by keys
+**************************************************************/
+
+_.mixin({
+  'sortKeysBy': function (obj, comparator) {
+    var keys = _.sortBy(_.keys(obj), function (key) {
+      return comparator ? comparator(obj[key], key) : key;
+    });
+
+    return _.object(keys, _.map(keys, function (key) {
+      return obj[key];
+    }));
+  }
+});
+
+/*************************************************************
 Global Time Vars
 **************************************************************/
 
@@ -21,6 +37,15 @@ sevenDaysAgo.setDate(sevenDaysAgo.getDate()-7)
 sevenDaysAgo.setHours(0)
 sevenDaysAgo.setMinutes(0)
 sevenDaysAgo.setSeconds(0)
+
+var yesterday = new Date();
+yesterday.setDate(yesterday.getDate()-2);
+
+var startOfYesterday = new Date();
+startOfYesterday.setDate(startOfYesterday.getDate()-2);
+startOfYesterday.setHours(0)
+startOfYesterday.setMinutes(0)
+startOfYesterday.setSeconds(0)
 
 /*************************************************************
 Mongoose Queries
@@ -38,7 +63,15 @@ exports.distinctStates = function(cb) {
 
 exports.distinctTrendsToday = function(cb) {
   USTrend.distinct("trend_name")
-         .where({created_at: {$gt: startofToday, $lt: today}})
+         .where({created_at: {$gt: startOfYesterday, $lt: yesterday}})
+         .exec(cb);
+}
+
+exports.cityCountTrendingYesterday = function(trendName, cb) {
+  USTrend.distinct("location_name")
+         .count()
+         .where({trend_name : trendName, 
+                 created_at: {$gt: startOfYesterday, $lt: yesterday}})
          .exec(cb);
 }
 
@@ -55,6 +88,56 @@ exports.statesTrendingThisWeek = function(trendName, cb) {
                  created_at: {$gt: sevenDaysAgo, $lt: today}})
          .exec(cb);
 }
+
+
+exports.usTrendsToday = function(cb) {
+  var trend_data = {};
+
+  //find all trends created today
+  exports.distinctTrendsToday(function(err, trends) {
+    if (err) {
+      console.log("error", err); 
+
+    } else {
+       // console.log("trends", trends);
+
+      async.each(trends, function(trend, next) {
+
+        //return count of which cities have that trend
+        exports.cityCountTrendingYesterday(trend, function(err, count) {
+          if (err) {
+            console.log("error", err); 
+
+          } else {
+            // console.log("count", count);
+
+            trend_data[trend] = count;
+            next();
+          }
+        });
+
+      }, function(err) {
+        if (err) {
+          console.log("A city failed to process.");
+
+        } else {
+
+          //sort by city count in descending order
+          var result =  _.sortKeysBy(trend_data, function (value, key) {
+            //changes from ascending to descending sort
+            return -(value);
+          });
+
+          cb(result);
+
+          // console.log("result", result);    
+        }
+      });
+    }
+  });
+}
+
+// exports.usTrendsToday();
